Optimized int32_t to int8_t; hopefully fixed OS X compile error; fixed wchar_t/char mismatch on windows (still only has been tested/compiled on linux)

+++ src/android/main.c
@@ -520,8 +520,8 @@ void config_osdefaults(UTOX_SAVE *r)
 
 static void android_main(struct android_app* state){
     _Bool theme_was_set_on_argv;
-    int32_t launch_at_startup;
-    int32_t set_show_window;
+    int8_t launch_at_startup;
+    int8_t set_show_window;
     _Bool no_updater;
     
     parseArgs(NULL, NULL, &theme_was_set_on_argv, &launch_at_startup, &set_show_window, &no_updater);
